Fixed Table Header WebViews.
This is somewhat of a hacky solution but I'm using
two webviews. One webview is the header info and the two
webview is the pure data.

Plus I refactor the code a bit to pass the dictionary to
the FullChartViewController instead of the raw html.

The fullchartviewcontroller generates the correctly html.

Also the Case Depth Chart is shown instead of going to
the selection picker screen like the hardness chart.